Change Y scale reference lever type to f64

+++ src/scpi/subsystems/display/window/trace/mod.rs
@@ -15,7 +15,7 @@ pub struct ScpiDisplayTrace {
 #[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub enum TraceCommand {
     GetYScaleReferenceLevel,
-    SetYScaleReferenceLevel(isize),
+    SetYScaleReferenceLevel(f64),
 }
 
 impl Display for ScpiDisplayTrace {
